+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # gridlinks
 
 Changes for version 0.0.9
+- added better tooltip rendering
+- added apache section
 - allow yaml as a data format
 - added gridlinks json schema for data